fixed imports in tests
